# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "BailFacile OpenApi Documentation",
        "description": "A Rest API dedicated to manage documents",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8001/api/",
            "description": "Demo API Server"
        }
    ],
    "paths": {
        "/documents": {
            "get": {
                "tags": [
                    "Documents"
                ],
                "summary": "Get a paginated list of documents",
                "description": "Returns list of paginated documents",
                "operationId": "getDocumentsLists",
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "query",
                        "description": "User owning the document",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "slug",
                        "in": "query",
                        "description": "Filter by document type slug",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "created_at",
                        "in": "query",
                        "description": "Filter document with date > created_at",
                        "required": false,
                        "schema": {
                            "type": "date"
                        }
                    },
                    {
                        "name": "update_at",
                        "in": "query",
                        "description": "Filter document with date > updated_at",
                        "required": false,
                        "schema": {
                            "type": "date"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "422": {
                        "description": "Wrong arguments"
                    }
                }
            },
            "post": {
                "tags": [
                    "Documents"
                ],
                "summary": "Create a document",
                "description": "Creates a document",
                "operationId": "postDocument",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "user_id",
                                    "slug"
                                ],
                                "properties": {
                                    "user_id": {
                                        "type": "integer",
                                        "format": "integer"
                                    },
                                    "slug": {
                                        "type": "string",
                                        "format": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Wrong parameters"
                    }
                }
            }
        },
        "/documents/{id}": {
            "get": {
                "tags": [
                    "Documents"
                ],
                "summary": "Get a document",
                "description": "Returns a document",
                "operationId": "getDocument",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Document id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "404": {
                        "description": "User does not exist"
                    }
                }
            },
            "put": {
                "tags": [
                    "Documents"
                ],
                "summary": "E-sign a document",
                "description": "E-sign a document",
                "operationId": "updateDocument",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Document id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Document is not e-sinable or already signed"
                    },
                    "404": {
                        "description": "Document does not exist"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Documents"
                ],
                "summary": "Delete existing document",
                "description": "Deletes a record and returns record",
                "operationId": "deleteDocument",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Document id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        }
    },
    "tags": [
        {
            "name": "Documents",
            "description": "API Endpoints of Documents"
        }
    ],
    "security": [
        []
    ]
}